renamed old cfg files
